requirements.txt
wordcloud
numpy
six
nltk
python-igraph
pandas
matplotlib
pillow
